+++ docs/source/history.rst
@@ -2,6 +2,7 @@ Release history
 ===============
 
 Version 0.4
+  - The encoders sample format is 's16le' except for the 'audio/l16' encoder.
   - The encoder command line is now updated with ``pa-dlna.conf`` user
     configuration.
   - Fix the parec command line length keeps increasing at each new track when